vlserver: Remove --enable-ubik-read-while-write
Commit c5b52c81 (vlserver: --enable-ubik-read-while-write configure
option) allowed the ubik "read-while-write"/"ReadAnyWrite"
functionality to be turned on with a configure switch.

There is no reason this needs to be a build-time switch; this makes it
very cumbersome to change the option, it forces the switch to be
global for all daemons, and makes it difficult for a user to see
whether the option is turned on or not.

Remove this build-time switch, so a runtime option can be added later.
Do not add such a runtime option in this commit yet, since the feature
is currently of uncertain stability. Future commits will add a runtime
option when there is more confidence in the functionality.

Change-Id: Ia3806eeefb7ee6330f7b9a8c048b83d0511df196
Reviewed-on: https://gerrit.openafs.org/16971
Reviewed-by: Michael Meffie <mmeffie@sinenomine.net>
Reviewed-by: Mark Vitale <mvitale@sinenomine.net>
Reviewed-by: Marcio Brito Barbosa <mbarbosa@sinenomine.net>
Reviewed-by: Andrew Deason <adeason@sinenomine.net>
Tested-by: BuildBot <buildbot@rampaginggeek.com>